Fix remaining duplicate schema errors in acts_as_taggable migrations

+++ db/migrate/20260509195110_add_tenant_to_taggings.acts_as_taggable_on_engine.rb
@@ -3,12 +3,13 @@
 # This migration comes from acts_as_taggable_on_engine (originally 7)
 class AddTenantToTaggings < ActiveRecord::Migration[6.0]
   def self.up
-    add_column ActsAsTaggableOn.taggings_table, :tenant, :string, limit: 128
+    # ADDED: if_not_exists: true
+    add_column ActsAsTaggableOn.taggings_table, :tenant, :string, limit: 128, if_not_exists: true
     add_index ActsAsTaggableOn.taggings_table, :tenant unless index_exists? ActsAsTaggableOn.taggings_table, :tenant
   end
 
   def self.down
-    remove_index ActsAsTaggableOn.taggings_table, :tenant
-    remove_column ActsAsTaggableOn.taggings_table, :tenant
+    remove_index ActsAsTaggableOn.taggings_table, :tenant if index_exists? ActsAsTaggableOn.taggings_table, :tenant
+    remove_column ActsAsTaggableOn.taggings_table, :tenant, if_exists: true
   end
 end
